feat(users): add Jellyfin personal rating endpoints
Remux stored no personal rating, so UserItemDataDto.Rating and .Likes were
always null and the /Rating endpoints returned 404.

Jellyfin stores this as a 0-10 double on UserItemData and derives Likes from
it at a 6.5 threshold, rather than keeping a separate boolean. Its /Rating
endpoint has historically been a shorthand over that double, writing 10 for a
like and 1 for a dislike, with DELETE clearing it.

Also takes an optional numeric `rating` alongside `likes`. When both are
given `rating` wins, and Likes stays derived, so the shape is additive and
existing clients are unaffected.

The range and the like threshold live on `db::UserRating`, a parsed newtype,
so an out-of-range or non-finite score cannot reach `UserMediaState.rating`.
NaN needs rejecting on its own: every comparison against it is false, so a
bare range check would pass it through. Likes is derived by
`UserMediaState::likes` rather than in the DTO mapping, and the write itself
is `UserMediaState::set_rating`, matching update_playback and the other state
mutations.

UserItemDataDto.Rating becomes f64, matching Jellyfin's double; the f32 it
was would have rounded a score across the like threshold.

Both the current /useritems/{id}/rating route and the legacy
/users/{user_id}/items/{id}/rating alias are included, since Jellyfin keeps
both and clients use either.

+++ crates/remux-sdks/src/remux/mod.rs
@@ -3010,7 +3010,9 @@ pub struct ProviderIds {
 
 #[dto]
 pub struct UserItemDataDto {
-    pub rating: Option<f32>,
+    /// Jellyfin's `UserItemData.Rating`: a 0-10 double, with `Likes` derived
+    /// from it rather than sent separately.
+    pub rating: Option<f64>,
     #[default(false)]
     pub played: bool,
     pub last_played_date: Option<DateTime<Utc>>,
